+++ Homie.API/appsettings.json
@@ -1,4 +1,7 @@
 {
+  "ConnectionStrings": {
+    "MainDatabaseConnection": "Server=DESKTOP-03UKV3M\\SQLEXPRESS; Database=HomieDb;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=False;ApplicationIntent=ReadWrite;MultiSubnetFailover=False"
+  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
