--- Homie.Data/HomieDbContext.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-public class HomieDbContext : DbContext
-{
-    public HomieDbContext(DbContextOptions<HomieDbContext> options) : base(options) { }
-
-    public HomieDbContext() { }
-}
