--- Homie.Data/Startup.cs
@@ -1,31 +1,0 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Homie.Data
-{
-    public class Startup
-    {
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        public IConfiguration Configuration { get; }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            IConfiguration configuration = new ConfigurationBuilder().SetBasePath(Directory.GetCurrentDirectory()).AddJsonFile("appsettings.json").Build();
-
-            string connectionString = configuration.GetConnectionString("MainDatabaseConnection");
-
-            services.AddDbContext<HomieDbContext>(options =>
-                options.UseSqlServer(connectionString));
-        }
-
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        { }
-    }
-}
